minor corrections
